Move Enum to cpp and add license files
